requirements.txt
asgiref==3.7.2
azure-core==1.28.0
azure-storage-blob==12.17.0
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
crispy-bootstrap4==2022.1
cryptography==41.0.3
Django==4.2.3
django-crispy-forms==2.0
django-storages==1.13.2
djangorestframework==3.14.0
idna==3.4
isodate==0.6.1
mypy-extensions==1.0.0
mysqlclient==2.1.1
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.9.1
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
whitenoise==6.5.0
